# foundation/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DeliverableSpec": {
        "properties": {
          "acceptance_criteria": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Acceptance Criteria",
            "type": "array"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "format": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Format",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "format",
          "acceptance_criteria"
        ],
        "title": "DeliverableSpec",
        "type": "object"
      },
...
      "ExecutionPackContent": {
        "properties": {
          "assumptions": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Assumptions",
            "type": "array"
          },
          "deliverables": {
            "items": {
              "$ref": "#/components/schemas/DeliverableSpec"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Deliverables",
            "type": "array"
          },
          "in_scope": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "In Scope",
            "type": "array"
          },
          "objective": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "out_of_scope": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Out Of Scope",
            "type": "array"
          },
          "problem_statement": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Problem Statement",
            "type": "string"
          },
          "recommended_agents": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Recommended Agents",
            "type": "array"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Risks",
            "type": "array"
          },
          "success_criteria": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Success Criteria",
            "type": "array"
          },
          "workstreams": {
            "items": {
              "$ref": "#/components/schemas/WorkstreamSpec"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Workstreams",
            "type": "array"
          }
        },
        "required": [
          "objective",
          "problem_statement",
          "success_criteria",
          "in_scope",
          "deliverables",
          "workstreams"
        ],
        "title": "ExecutionPackContent",
        "type": "object"
      },
      "ExecutionPackGenerateRequest": {
        "properties": {
          "change_summary": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Summary"
          }
        },
        "title": "ExecutionPackGenerateRequest",
        "type": "object"
      },
      "ExecutionPackResponse": {
        "properties": {
          "current_version": {
            "$ref": "#/components/schemas/ExecutionPackVersionResponse"
          },
          "current_version_number": {
            "title": "Current Version Number",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ExecutionPackVersionResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "current_version_number",
          "current_version",
          "versions"
        ],
        "title": "ExecutionPackResponse",
        "type": "object"
      },
      "ExecutionPackVersionResponse": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "change_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Summary"
          },
          "content": {
            "$ref": "#/components/schemas/ExecutionPackContent"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version_number",
          "content",
          "change_summary",
          "created_at",
          "approved_at"
        ],
        "title": "ExecutionPackVersionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkstreamSpec": {
        "properties": {
          "goal": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "name",
          "goal",
          "steps"
        ],
        "title": "WorkstreamSpec",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Governed execution API for Ryan Agent OS.",
    "title": "Ryan Agent OS API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/execution-pack": {
      "get": {
        "operationId": "get_execution_pack_route_api_v1_projects__project_id__execution_pack_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Execution Pack Route",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "generate_execution_pack_route_api_v1_projects__project_id__execution_pack_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionPackGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Execution Pack Route",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}